+++ LuccaDevises/ExplorationPath.cs
@@ -6,26 +6,26 @@ namespace LuccaDevises
     public class ExplorationPath
     {
 
-        private Dictionary<string, float> Path;
+        private Dictionary<string, double> Path;
         public bool IsValidPath { get; set; }
 
         public ExplorationPath()
         {
-            Path = new Dictionary<string, float>();
+            Path = new Dictionary<string, double>();
             IsValidPath = false;
         }
 
-        public void AddToPath(CurrencyNode node, float value)
+        public void AddToPath(CurrencyNode node, double value)
         {
             Path.Add(node.Name, value);
         }
 
-        public void AddToPath(string nodeName, float value)
+        public void AddToPath(string nodeName, double value)
         {
             Path.Add(nodeName, value);
         }
 
-        public Dictionary<string, float> GetPath()
+        public Dictionary<string, double> GetPath()
         {
             return Path;
         }
@@ -33,10 +33,10 @@ namespace LuccaDevises
         public ExplorationPath Clone()
         {
             ExplorationPath copy = new ExplorationPath();
-            copy.Path = new Dictionary<string, float>();
+            copy.Path = new Dictionary<string, double>();
             copy.IsValidPath = IsValidPath;
 
-            foreach(KeyValuePair<string, float> keyValuePair in Path)
+            foreach(KeyValuePair<string, double> keyValuePair in Path)
             {
                 copy.Path.Add(keyValuePair.Key, keyValuePair.Value);
             }
